% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avs.r
\name{avs}
\alias{avs}
\title{Analisis Varians (ANOVA) dan Uji Lanjut (Post-hoc Test)}
\usage{
avs(
  file_jalur,
  file_kolom,
  lembar_nama,
  kolom_respon,
  jenis_rancangan,
  kolom_perlakuan,
  kolom_kelompok,
  jenis_ujilanjut
)
}
\arguments{
\item{file_jalur}{Jalur ke file data excel}

\item{file_kolom}{Tipe kolom data excel}

\item{lembar_nama}{Nama sheet yang akan digunakan}

\item{kolom_respon}{Nama kolom hasil dalam sheet yang digunakan}

\item{jenis_rancangan}{Tipe ANOVA ("RAK" atau "RAL")}

\item{kolom_perlakuan}{Nama kolom perlakuan}

\item{kolom_kelompok}{Nama kolom ulangan}

\item{jenis_ujilanjut}{Tipe uji lanjut ("BNT", "BNJ", "DMRT", atau "ScottKnott")}
}
\value{
Hasil ANOVA dan Uji Lanjut
}
\description{
Analisis Varians (ANOVA) dan Uji Lanjut (Post-hoc Test)
}
